Refactor the big file
